+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 8698046 - 2017-11-10: Fixed search issue with Newpct torrent provider Added subtitles category to Newpct torrent provider Quality of Newpct releases determined from title and url now 
+- * 52047a1 - 2017-11-10: Unidecode Newpct titles to avoid unicode issues during name parsing Refactored TNTVillage torrent provider Refactored regular regexes 
+- * 1c854a2 - 2017-11-10: Fixed search issue with Newpct torrent provider Added subtitles category to Newpct torrent provider Quality of Newpct releases determined from title and url now 
 - * 022e4f9 - 2017-11-10: Fixed unicode decode issues for EliteTorrent provider 
 - * 24e1037 - 2017-11-10: Release v9.1.70 
 - * 9498449 - 2017-11-10: Fixed typo in code for daily searcher